Win32: Disable fb transparency when it is broken
On Windows 7, when the colorization color is opaque (i.e. when
the Transparency setting is disabled under Personalization > Color), the
contents of the framebuffer is combined with the last frame using
additive blending instead of replacing the previous contents.

This commit enables framebuffer transparency only when colorization
color opacity is disabled, removing the previous workaround of setting
a layered window color key that led to rendering artifacts.

The colorization color opacity cannot be relied on alone, as it is
always true on Windows 10.

Fixes #1512.